+++ src/resources/documentResource.js
@@ -6,7 +6,7 @@ import {
   revertRowInListResource,
 } from './listResource'
 import { getLocal, saveLocal, deleteLocal } from './local'
-import { onDocUpdate } from './realtime'
+import { onDocUpdate, readSocket } from './realtime'
 import { getConfig } from '../utils/config'
 
 let documentCache = reactive({})
@@ -266,8 +266,9 @@ export function createDocumentResource(options, vm) {
     return doc
   }
 
-  if (options.realtime && vm.$socket) {
-    onDocUpdate(vm.$socket, out.doctype, (name) => {
+  let socket = options.realtime ? readSocket(vm) : undefined
+  if (socket) {
+    onDocUpdate(socket, out.doctype, (name) => {
       if (name == out.name) {
         out.get.fetch()
       }